# Speed up iteration over Variables objects