--- a/releasenotes/notes/feature-speed-up-variables-iteration-bfe9d6596e24572f.yaml
+++ b/releasenotes/notes/feature-speed-up-variables-iteration-bfe9d6596e24572f.yaml
@@ -1,0 +1,3 @@
+---
+features:
+  - Improve the performance of iteration over ``Variables`` objects.