New timing results for `for v in variables: pass`
```bash
                labels              
              --------- ------------
                string   34.3±0.7μs 
                index    11.2±0.3μs 
               integer   34.1±0.6μs 
```
vs the old timing results
```bash
                labels              
              --------- ------------
                string   63.0±0.3μs 
                index     53.1±1μs  
               integer    62.2±3μs  
```