@arcondello 2022-04-25
In this case the speed comes from implementing `.__iter__()` "by hand" rather than using the one inherited from `collections.abc.Sequence` which looks like
```python
    def __iter__(self):
        i = 0
        try:
            while True:
                v = self[i]
                yield v
                i += 1
        except IndexError:
            return
```
([source code](https://github.com/python/cpython/blob/971343eb569a3418aa9a0bad9b638cccf1470ef8/Lib/_collections_abc.py#L1034)).

By doing it "by hand" we can improve the performance in the range case, and even in the case with labels we can bypass a bunch of function calls.